@@ -29,6 +29,7 @@
       }
     ],
     "removePackageScripts": false,
+    "generateUpdatesFilesForAllChannels": true,
     "directories": {
       "app": "dist/server",
       "buildResources": "dist/server/assets/icons/build",
@@ -77,22 +78,22 @@
     "@angular-devkit/build-ng-packagr": "0.10.6",
     "@angular-devkit/core": "0.8.7",
     "@angular-devkit/schematics": "0.8.7",
-    "@angular/animations": "7.0.4",
+    "@angular/animations": "7.1.0",
     "@angular/cdk": "7.1.0",
     "@angular/cli": "7.0.6",
-    "@angular/common": "7.0.4",
-    "@angular/compiler": "7.0.4",
-    "@angular/compiler-cli": "7.0.4",
-    "@angular/core": "7.0.4",
+    "@angular/common": "7.1.0",
+    "@angular/compiler": "7.1.0",
+    "@angular/compiler-cli": "7.1.0",
+    "@angular/core": "7.1.0",
     "@angular/flex-layout": "7.0.0-beta.19",
-    "@angular/forms": "7.0.4",
-    "@angular/language-service": "7.0.4",
+    "@angular/forms": "7.1.0",
+    "@angular/language-service": "7.1.0",
     "@angular/material": "7.1.0",
-    "@angular/platform-browser": "7.0.4",
-    "@angular/platform-browser-dynamic": "7.0.4",
-    "@angular/router": "7.0.4",
-    "@nrwl/angular-console-enterprise-electron": "0.0.17",
-    "@nrwl/angular-console-enterprise-frontend": "0.0.17",
+    "@angular/platform-browser": "7.1.0",
+    "@angular/platform-browser-dynamic": "7.1.0",
+    "@angular/router": "7.1.0",
+    "@nrwl/angular-console-enterprise-electron": "0.0.31",
+    "@nrwl/angular-console-enterprise-frontend": "0.0.31",
     "@nrwl/schematics": "7.0.2",
     "@types/express": "4.16.0",
     "@types/graphql": "0.13.4",
@@ -105,8 +106,8 @@
     "angular-playground": "5.2.0",
     "apollo-angular": "1.5.0",
     "apollo-angular-link-http": "1.4.0",
-    "apollo-cache-inmemory": "1.3.10",
-    "apollo-client": "2.4.6",
+    "apollo-cache-inmemory": "1.3.11",
+    "apollo-client": "2.4.7",
     "apollo-link": "1.2.4",
     "apollo-link-error": "1.1.2",
     "apollo-server-express": "2.2.2",
@@ -153,7 +154,7 @@
     "rxjs-compat": "6.3.3",
     "shelljs": "0.8.2",
     "strip-json-comments": "2.0.1",
-    "ts-jest": "23.10.4",
+    "ts-jest": "23.10.5",
     "ts-node": "7.0.1",
     "tsickle": "0.34.0",
     "tslib": "1.9.3",
